faz muitacoisa

+++ controllers/MemberController.php
@@ -85,6 +85,6 @@ $postActions = array('create', 'update', 'changePassword', 'recoverPassword');
 if (isset($_POST['action']) && in_array($_POST['action'], $postActions)) {
     $action = $_POST['action'];
     MemberController::$action();
-} elseif (!empty($_GET)) {
+} elseif (!empty(key($_GET)) &&  key($_GET) == "delete") {
     MemberController::delete();
 }
