+++ controllers/PointController.php
@@ -84,6 +84,6 @@ $postActions = array('create', 'update', 'changePassword', 'recoverPassword');
 if (isset($_POST['action']) && in_array($_POST['action'], $postActions)) {
     $action = $_POST['action'];
     PointController::$action();
-} elseif (!empty($_GET)) {
+} elseif (!empty(key($_GET)) &&  key($_GET) == "delete") {
     PointController::delete();
 }
