+++ controllers/ScheduleController.php
@@ -9,7 +9,7 @@ class ScheduleController{
     public static function create() {
         if($_POST['ponto'] != ''){
 
-            $member = new Schedule($_POST['ponto']);
+            $member = new Schedule($_POST);
             try {
                 $member->create();
                 $_SESSION['msg'] = null;
@@ -81,6 +81,6 @@ $postActions = array('create', 'update', 'changePassword', 'recoverPassword');
 if (isset($_POST['action']) && in_array($_POST['action'], $postActions)) {
     $action = $_POST['action'];
     UserController::$action();
-} elseif (!empty($_GET)) {
+} elseif (!empty(key($_GET)) &&  key($_GET) == "delete") {
     UserController::delete();
 }
